requirements.txt
opencv-python
Flask
matplotlib
Keras
tqdm
numpy
sklearn
scikit-learn
imutils
pillow